Added method to post a plant

+++ Plant.cs
@@ -3,7 +3,7 @@ public class Plant
 {
 	public string? Species { get; set; }
 	public int LightNeeds { get; set; }
-    public double AskingPrice { get; set; }
+    public decimal AskingPrice { get; set; }
     public string? City { get; set; }
     public string? ZIP { get; set; }
     public bool Sold { get; set; }
